+++ tko_project_task_actions/models/__init__.py
@@ -22,4 +22,6 @@
 #
 ##############################################################################
 
+
+import task_action
 import project_task
