+++ tko_project_task_actions/views/configure_tasks_in_stages_view.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<openerp>
+    <data noupdate="1">
+
+
+        <function model="project.task" name="_set_tasks_in_stages"/>
+
+    </data>
+</openerp>
