+++ tko_project_task_actions/views/project_task_view.xml
@@ -1,6 +1,17 @@
 <openerp>
     <data>
 
+        <record id="task_type_edit" model="ir.ui.view">
+            <field name="name">project.task.type</field>
+            <field name="model">project.task.type</field>
+            <field name="inherit_id" ref="project.task_type_edit"/>
+            <field name="arch" type="xml">
+                <field name="sequence" position="after">
+                    <field name="task_ids" invisible="1"/>
+                </field>
+            </field>
+        </record>
+
         <record id="view_task_form_inherit" model="ir.ui.view">
             <field name="name">project.task.form</field>
             <field name="model">project.task</field>
